bubble

+++ 0-bubble_sort.c
@@ -10,7 +10,7 @@
 void bubble_sort(int *array, size_t size)
 {
 	size_t i, j;
-	int  temp;
+	size_t  temp;
 
 	if (array == NULL || size < 2)
 	return;
